* moved Redux Provider inside SemanticScreen component (removed from index.tsx) * moved all logic from SemanticScreen.tsx into a new component SemanticScreenLogic.tsx * deleted onAuthorUpdate (was never used) * ran prettier
